Move exceptions and the step decorator out of goth/runner/__init__.py
