+++ goth/runner/exceptions.py
@@ -23,5 +23,23 @@ class StopThreadException(Exception):
     """Exception used to stop a running `StoppableThread`."""
 
 
+class TestFailure(Exception):
+    """Base class for errors raised by the test runner when a test fails."""
+
+
+class StepTimeoutError(TestFailure):
+    """Raised on test step timeout."""
+
+    def __init__(self, step: str, time: float):
+        super().__init__(f"Step '{step}' timed out after {time:.1f} s")
+
+
+class TemporalAssertionError(TestFailure):
+    """Raised on temporal assertion failure."""
+
+    def __init__(self, assertion: str):
+        super().__init__(f"Temporal assertion '{assertion}' failed")
+
+
 class TimeoutError(Exception):
     """Exception for when a timeout occurs."""
